+++ CHANGELOG.md
@@ -1,13 +1,11 @@
-### 0.7.1 (Next)
+### 0.7.1 (2026/01/28)
 
 #### Features
 
-* Your contribution here.
 * [#91](https://github.com/ruby-grape/grape-swagger-entity/pull/91): Improve README and add documentation structure - [@numbata](https://github.com/numbata).
 
 #### Fixes
 
-* Your contribution here.
 * [#90](https://github.com/ruby-grape/grape-swagger-entity/pull/90): Fix description not rendered for non-array entity references - [@numbata](https://github.com/numbata).
 * [#86](https://github.com/ruby-grape/grape-swagger-entity/pull/86): Fix Array[Object] documentation type parsing - [@numbata](https://github.com/numbata).
 * [#87](https://github.com/ruby-grape/grape-swagger-entity/pull/87): Remove hidden attributes from required - [@bogdan](https://github.com/bogdan).